requirements.txt
python-firebase==1.2
Flask==0.10.1
Flask-OAuth==0.12
Flask-OAuthlib==0.9.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
httplib2==0.9.1
itsdangerous==0.24
oauth2==1.5.211
oauthlib==0.7.2
python-twitter==2.2
requests==2.7.0
requests-oauthlib==0.5.0
wsgiref==0.1.2
